applications/Copilot/requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
apispec==6.4.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.1
azure-core-tracing-opentelemetry==1.0.0b11
azure-identity==1.17.1
azure-monitor-opentelemetry==1.6.1
azure-monitor-opentelemetry-exporter==1.0.0b28
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dataclasses-json==0.6.4
deepdiff==6.7.1
Deprecated==1.2.14
deptry==0.16.1
distro==1.9.0
exceptiongroup==1.2.0
fixedint==0.1.6
flasgger==0.9.7.1
Flask==3.0.3
Flask-RESTful==0.3.10
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.11.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.50
langchain-openai==0.1.6
langchain-text-splitters==0.0.2
langsmith==0.1.65
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow_dataclass==8.7.0
microsoft-kiota-abstractions==1.3.3
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.2
microsoft-kiota-serialization-form==0.1.0
microsoft-kiota-serialization-json==1.3.0
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
mistune==3.0.2
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.25.2
opentelemetry-api==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-dbapi==0.47b0
opentelemetry-instrumentation-django==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-flask==0.47b0
opentelemetry-instrumentation-psycopg2==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-urllib==0.47b0
opentelemetry-instrumentation-urllib3==0.47b0
opentelemetry-instrumentation-wsgi==0.47b0
opentelemetry-resource-detector-azure==0.1.4
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.9.15
packaging==23.2
pendulum==3.0.0
portalocker==2.8.2
psutil==5.9.8
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pypdf==4.2.0
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
std-uritemplate==1.0.5
tenacity==8.3.0
tiktoken==0.7.0
time-machine==2.14.2
tk==0.1.0
tk64==0.1
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
typeguard==4.0.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
